feat: Implement Master Backplane Architecture for enhanced configuration management and atomic updates
